doc/guide: clean up the chapter on creating packages
Url2pkg is no longer interactive, it just takes the URL as command line
argument.  It is not necessary to create the package directory manually,
as url2pkg does that as well.

Since Perl packages are handled by url2pkg (just like Python packages),
there is no need to list them in an effectively empty section.