py-cmudict: add new package, version 1.0.2
CMUdict is a versioned Python wrapper package for the CMU Pronouncing
Dictionary data files. The main purpose is to expose the data with little
or no assumption on how it is to be used.

Similar functionality is also provided by nltk. My motivations for
importing this package are that (1) nltk doesn't appear to be keeping
its packaging of the dictionary data up to date, and (2) this has no
dependency chain beyond Python itself.