py-borgbackup: remove ALTERNATIVES
The current ALTERNATIVES/post-install was not enough to make this package
avoid a self-conflict (due to the man pages). Since this is an application
that which no other packages depend upon, there is no need to have it
installed for multiple python versions at the same time, so simplify
this.

Bump PKGREVISION.