nim: Update to 1.6.4
Changelog:
Version 1.6.4 released

08 February 2022 The Nim Team

The Nim team is happy to announce version 1.6.4, our second patch release for
Nim 1.6.

Version 1.6.4 is a result of a month and a half of hard work, and it contains
33 commits, fixing one major regression and bringing some general improvements
over 1.6.2.

The most important fix is for the C FFI regression introduced in 1.6.2.

We would recommend to all of our users to upgrade and use version 1.6.4.

Bugfixes

These reported issues were fixed:

  * Fixed "Potential C FFI regression" (#19342)
  * Fixed "re.split unexpected results with zero-width characters" (#14468)
  * Fixed "Out-of-bounds in strformat" (#19107)
  * Fixed "Adding an empty list to a non-empty list breaks the latter list" (
    #19297)
  * Fixed "Wrong result when using varargs with var arguments." (#16617)
  * Fixed "Adding an empty DoublyLinkedList to a non-empty DoublyLinkedList
    breaks the latter list" (#19314)
  * Fixed "Silent FFI bug when passing array inside object using gc:orc" (#
    19497)