lang/guile18: import guile18-1.8.8nb11
Re-import of lang/guile to avoid having an ancient version as the default.

GUILE, GNU's Ubiquitous Intelligent Language for Extension, is a library
that implements the Scheme language plus various convenient facilities.
It's designed so that you can link it into an application or utility to
make it extensible. Our plan is to link this library into all GNU programs
that call for extensibility.

This is an extremely old version of guile.  It is installed into
guile/1.8 within PREFIX.