py-hg-evolve: update to 10.5.0.
This is a feature release. The most notable changes are: a new
command:  fixup, which is used to add changes from the working
directory to an arbitrary revision. An issue with histedit not
preserving topics in certain situations was also fixed, see
https://bz.mercurial-scm.org/show_bug.cgi?id=6550. Evolve is now
compatible with the upcoming Mercurial 6.1, which includes head
computation code that is obsolescence-aware (adapted from the evolve
extension). With an up-to-date client and server you should expect
hg push to take much less time. Also compatibility with Mercurial
4.7 was dropped in this release.