py-notify2: add new package, version 0.3.1
A modern Python interface to DBus notifications, supporting Python 3.
This effectively replaces what was once called "py-notify" in pkgsrc,
and then became "py-notify-python". It is actually required by
print/hplip, if the qt5 option is enabled (which I am working on
fixing).

(There is also a completely unrelated package as sysutils/py-notify at
present, co-incidentally also at version 0.3.1, which apparently was
mistakenly imported overtop the original py-notify, but is not the same
package at all: it should have been placed in devel, not sysutils, as
it doesn't deal with notifications integration but rather with general
development abstractions.)

Originally packaged in wip by K.I.A. Derouiche. Minor adjustments for
pkgsrc and a correction (missing py-dbus dependency) by me.