Add opencpn-plugin-plots version 2.2.12
OpenCPN is a free software (GPLv2) project to create a concise chart plotter
and navigation software, for use underway or as a planning tool. OpenCPN is
developed by a team of active sailors using real world conditions for
program testing and refinement.

This plugin is designed to monitor speed and course to make the changes to
sail sweep obvious. It may be used to better understand the sailing
characteristics of a particular sailing vessel, or for fine tuning to give
optimal results.