py-music21: update to 7.3.1
The full upstream change log can't be condensed easily, highlights are:

Version 7.3 is a bug-fix, improvement, and maintenance release of music21. Aside from a few obscure corners of the code, it is designed to be fully backwards compatible with the previous v7 releases (7.1). It represents half a year's work since the original v7.1 release.

Thanks go especially to Jacob Tyler Walls who really steered the contributions to this release.

Among the major improvements in this branch are:

    Full support for Python 3.10.
        I've been doing all my recent dev on 3.10 taking advantage of new features, and it's great, and plays well with numpy, tensorflow, and so many other cutting edge analysis tools.
        In keeping with music21's policy of supporting the last three versions of Python, music21 v7.3 will be the last version of music21 to support Python 3.7. A decision on whether 3.8 or 3.9 will be the minimum version for music21 v8 will come later, depending on its release date; we may also update policy somewhat and raise the minimum version to 3.9 during the v8 cycle.

(7.3.0 had an installation bug -- replacing with 7.3.1)

For full details, please refer to
https://github.com/cuthbertLab/music21/releases/tag/v7.3.0